refactor(auth): extract shared auth user hook
