+++ frontend/src/components/AuthButton.tsx
@@ -1,36 +1,9 @@
-import { useEffect, useState } from 'react';
-import type { User } from '@supabase/supabase-js';
-import { supabase } from '../../lib/supabaseClient';
 import { LoginButton } from './LoginButton';
 import { LogoutButton } from './LogoutButton';
+import { useAuthUser } from '../hooks/useAuthUser';
 
 export const AuthButton = () => {
-  const [user, setUser] = useState<User | null>(null);
-  const [isCheckingAuth, setIsCheckingAuth] = useState(true);
-
-  useEffect(() => {
-    const getUser = async () => {
-      const {
-        data: { user },
-      } = await supabase.auth.getUser();
-
-      setUser(user);
-      setIsCheckingAuth(false);
-    };
-
-    getUser();
-
-    const {
-      data: { subscription },
-    } = supabase.auth.onAuthStateChange((_event, session) => {
-      setUser(session?.user ?? null);
-      setIsCheckingAuth(false);
-    });
-
-    return () => {
-      subscription.unsubscribe();
-    };
-  }, []);
+  const { user, isCheckingAuth } = useAuthUser();
 
   if (isCheckingAuth) {
     return <p>Checking auth...</p>;
